chore: move language service in engine
